Move rust into namespace
